Api changes for search and edition controller addtional field

+++ database/migrations/2016_05_03_125355_update_ads_column4_table.php
@@ -13,11 +13,14 @@ class UpdateAdsColumn4Table extends Migration
     public function up()
     {
         Schema::table('ads', function ($table) {
-            $table->string('id_image')->after('ad_type');
-            //Foreign Keys
-            $table->foreign('id_image')
-                    ->references('id_image')->on('archive_images')
-                    ->onDelete('cascade');
+            if (!Schema::hasColumn('ads', 'id_image'))
+            {
+                $table->string('id_image')->after('ad_type');
+                //Foreign Keys
+                $table->foreign('id_image')
+                        ->references('id_image')->on('archive_images')
+                        ->onDelete('cascade');
+            }    
         });
     }
